Divide into HTML5toTouch file and Change to setQueryData
